Maggot Slime Potion
I thought it might be fun to entertain the idea of a maggot slime potion.
Hope you all like! 😄

----
@WildX : Seems like a good vendor item for Hurnscald general goods and maybe
a rare drop from some humanoid mob, I'd like to keep manmade things like
potions as items that are dropped by humanoids since it makes little sense
for a maggot to carry a bottled potion with it.